+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Update Restic 0.18.1 (built with Go 1.24)
 * Update Python 3.14 and dependencies
 * Update base Docker image to Alpine 3.23
+* Include Restic Exporter version in the log traces
 * Use UV as package manager for local development
 * Cleanup the code and add unit tests
 * Add warning in the readme about remote Restic repo costs